Move docker common files into a shared foldergs
