+++ src/common/docker-compose/index.ts
@@ -7,10 +7,10 @@ import os from 'os';
 import pLimit from 'p-limit';
 import path from 'path';
 import untildify from 'untildify';
-import which from 'which';
 import { ArchitectError, ComponentNode, DependencyGraph, Dictionary, GatewayNode, IngressEdge, ResourceSlugUtils, ServiceNode, TaskNode } from '../../';
 import LocalPaths from '../../paths';
-import { DockerHelper, restart } from '../utils/docker';
+import { DockerHelper } from '../docker/helper';
+import { restart } from '../docker/cmd';
 import PortUtil from '../utils/port';
 import { DockerComposeProject, DockerComposeProjectWithConfig } from './project';
 import DockerComposeTemplate, { DockerService } from './template';
