+++ src/common/docker/helper.ts
@@ -1,21 +1,7 @@
 import execa, { Options } from 'execa';
 import which from 'which';
+import { docker } from './cmd';
 
-export const docker = async (args: string[], opts = { stdout: true }, execa_opts?: Options): Promise<any> => {
-  if (process.env.TEST === '1') {
-    return;
-  }
-  const cmd = execa('docker', args, execa_opts);
-  if (opts.stdout) {
-    cmd.stdout?.pipe(process.stdout);
-    cmd.stderr?.pipe(process.stderr);
-  }
-  return await cmd;
-};
-
-export const restart = async (container_id: string): Promise<void> => {
-  return docker(['restart', container_id]);
-};
 
 /**
  * this method splits the tag off of an image string. this logic is not straightforward as the image string may contain a port.
