+++ test/commands/task.test.ts
@@ -2,7 +2,7 @@ import { expect } from 'chai';
 import path from 'path';
 import sinon from 'sinon';
 import { DockerComposeUtils } from '../../src/common/docker-compose/index';
-import * as Docker from '../../src/common/utils/docker';
+import * as Docker from '../../src/common/docker/cmd';
 import { ComponentSlugUtils, ComponentVersionSlugUtils, resourceRefToNodeRef, ResourceSlugUtils } from '../../src/';
 import { mockArchitectAuth, MOCK_API_HOST } from '../utils/mocks';
