+++ test/utils/mocks.ts
@@ -3,8 +3,8 @@ import path from 'path';
 import AuthClient from '../../src/app-config/auth';
 import SecretUtils from '../../src/architect/secret/secret.utils';
 import { DockerComposeUtils } from '../../src/common/docker-compose';
-import * as Docker from '../../src/common/utils/docker';
-import DockerBuildXUtils from '../../src/common/utils/docker-buildx.utils';
+import * as Docker from '../../src/common/docker/cmd';
+import DockerBuildXUtils from '../../src/common/docker/buildx.utils';
 
 export const MOCK_API_HOST = 'http://mock.api.localhost';
 export const MOCK_REGISTRY_HOST = 'http://mock.registry.localhost';
